pyproject.toml
[tool.poetry]
name = "gd-backend"
version = "0.1.0"
description = ""
authors = ["Kamil Bobola <kamilb9610@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
environs = "^11.2.1"
django-filter = "^24.3"
django-rest-knox = "^5.0.2"
geopy = "^2.4.1"
pydantic = "^2.10.3"
redis = "^5.2.1"
hiredis = "^3.1.0"
pillow = "^11.0.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}
django-storages = "^1.14.4"
django-compressor = "^4.5.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
markdown = "^3.7"
pylint = "^3.3.2"
isort = "^5.13.2"
pre-commit = "^4.0.1"
django-zeal = "^2.0.2"


[tool.poetry.group.black.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
